+++ src/test/resources/data/sql/h2/sbp-h2.sql
@@ -129,7 +129,7 @@ create table tx
     time   time                              not null,
     descr  varchar(100)                      not null,
     type   char(1)                           not null,
-    amount decimal                           not null,
+    amount double                            not null,
     constraint tx_ref_unq unique (ref)
 );
